requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.9
